backend/requirements.txt
flask
flask-cors
apscheduler
psycopg2-binary
python-dotenv
requests
gradio
gunicorn
openai
